# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de prueba NITSNETS",
        "description": "Esta es una primera versión de la API",
        "contact": "jorgepastor9120@gmail.com",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Mostrar el listado de reservas",
                "description": "Mostramos el listado de los registros solicitados.",
                "operationId": "daf1d30d180ba251d3a594b8cc06fcc5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Registra una reserva",
                "description": "Registrar una pista.",
                "operationId": "f6f5bf7dc5def23c323cebb98b5d1e5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "court_id",
                                    "date",
                                    "hour_reserve_id"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "4"
                                    },
                                    "court_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "7"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-25"
                                    },
                                    "hour_reserve_id": {
                                        "description": "Hora del día (formato de 24 horas desde las 08:00 a 22:00)",
                                        "type": "integer",
                                        "format": "int20",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5,
                                            6,
                                            7,
                                            8,
                                            9,
                                            10,
                                            11,
                                            12,
                                            13
                                        ],
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "201": {
                        "description": "La pista ha sido creado"
                    },
                    "404": {
                        "description": "No se ha podido registrar la pista."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{id}": {
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Actualiza una reserva",
                "description": "Actualizar un registro.",
                "operationId": "09997977865b4eac7a19d9a76304b362",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Actualiza una reserva",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "court_id",
                                    "date",
                                    "hour_reserve_id"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "4"
                                    },
                                    "court_id": {
                                        "type": "integer",
                                        "format": "int20",
                                        "example": "7"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-25"
                                    },
                                    "hour_reserve_id": {
                                        "description": "Hora del día (formato de 24 horas desde las 08:00 a 22:00)",
                                        "type": "integer",
                                        "format": "int20",
                                        "enum": [
                                            1,
                                            2,
                                            3,
                                            4,
                                            5,
                                            6,
                                            7,
                                            8,
                                            9,
                                            10,
                                            11,
                                            12,
                                            13
                                        ],
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido actualizar el registro."
                    },
                    "405": {
                        "description": "No se ha permitido actualizar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Elimina una reserva",
                "description": "Elimina un registro.",
                "operationId": "6d30d72eafb0ab89647f6058bd75774c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int20"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Elimina una reserva"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No se ha podido eliminar el registro."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/bookings/{date}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Mostrar el listado de reservas según una fecha",
                "description": "Mostramos el listado de los registros solicitados según una fecha.",
                "operationId": "cb632dc7d91dd90d22e42ec0877277fe",
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2023-10-25"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No se ha podido cargar el listado."
                    },
                    "503": {
                        "description": "El servidor no está disponible en este momento"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "name": "authorization",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/api/v1/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}